safari-600.7-branch/LayoutTests/svg/custom/svg-parse-overflow-1-expected.txt

CONSOLE MESSAGE: line 9: Error: Invalid value for <foreignObject> attribute y="105000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000"
